fixed view and refactored organization file
