--- frontend/src/pages/Organization/Columns/scanColumns.tsx
@@ -1,99 +1,0 @@
-import React from 'react';
-import * as authContextOrg from '../authContextOrg';
-import { Button } from '@mui/material';
-import { 
-  Organization as OrganizationType,
-  Scan,
-  ScanSchema
-} from 'types';
-import { Column } from 'react-table';
-
-
-const apiPost = authContextOrg.apiPost;
-const setFeedbackMessage = authContextOrg.setFeedbackMessage;
-const [ organization, setOrganization ] = React.useState<OrganizationType>();
-const [scans] = React.useState<Scan[]>([]);
-const [scanSchema] = React.useState<ScanSchema>({});
-
-
-
-export const updateScan = async (scan: Scan, enabled: boolean) => {
-    try {
-      if (!organization) return;
-      await apiPost(
-        `/organizations/${organization?.id}/granularScans/${scan.id}/update`,
-        {
-          body: {
-            enabled
-          }
-        }
-      );
-      setOrganization({
-        ...organization,
-        granularScans: enabled
-          ? organization.granularScans.concat([scan])
-          : organization.granularScans.filter(
-              (granularScan) => granularScan.id !== scan.id
-            )
-      });
-    } catch (e: any) {
-      setFeedbackMessage({
-        message:
-          e.status === 422 ? 'Error updating scan' : e.message ?? e.toString(),
-        type: 'error'
-      });
-      console.error(e);
-    }
-  };
-
-export const scanColumns: Column<Scan>[] = [
-  
-    {
-      Header: 'Name',
-      accessor: 'name',
-      width: 150,
-      id: 'name',
-      disableFilters: true
-    },
-    {
-      Header: 'Description',
-      accessor: ({ name }) => scanSchema[name] && scanSchema[name].description,
-      width: 200,
-      minWidth: 200,
-      id: 'description',
-      disableFilters: true
-    },
-    {
-      Header: 'Mode',
-      accessor: ({ name }) =>
-        scanSchema[name] && scanSchema[name].isPassive ? 'Passive' : 'Active',
-      width: 150,
-      minWidth: 150,
-      id: 'mode',
-      disableFilters: true
-    },
-    {
-      Header: 'Action',
-      id: 'action',
-      maxWidth: 100,
-      Cell: ({ row }: { row: { index: number } }) => {
-        if (!organization) return null;
-        const enabled = organization.granularScans.find(
-          (scan) => scan.id === scans[row.index].id
-        );
-        return (
-          <Button
-            type="button"
-            onClick={() => {
-              updateScan(scans[row.index], !enabled);
-            }}
-          >
-            {enabled ? 'Disable' : 'Enable'}
-          </Button>
-        );
-      },
-      disableFilters: true
-    }
-  ];
-
-
