--- frontend/src/pages/Organization/Columns/scanTaskColumns.ts
@@ -1,62 +1,0 @@
-import { Column } from 'react-table';
-import { ScanTask } from 'types';
-import { formatDistanceToNow, parseISO } from 'date-fns';
-
-const dateAccessor = (date?: string) => {
-  return !date || new Date(date).getTime() === new Date(0).getTime()
-      ? 'None'
-      : `${formatDistanceToNow(parseISO(date))} ago`;
-};
-
-
-export  const scanTaskColumns: Column<ScanTask>[] = [
-    {
-      Header: 'ID',
-      accessor: 'id',
-      disableFilters: true
-    },
-    {
-      Header: 'Status',
-      accessor: 'status',
-      disableFilters: true
-    },
-    {
-      Header: 'Type',
-      accessor: 'type',
-      disableFilters: true
-    },
-    {
-      Header: 'Name',
-      accessor: ({ scan }) => scan?.name,
-      disableFilters: true
-    },
-    {
-      Header: 'Created At',
-      accessor: ({ createdAt }) => dateAccessor(createdAt),
-      disableFilters: true,
-      disableSortBy: true
-    },
-    {
-      Header: 'Requested At',
-      accessor: ({ requestedAt }) => dateAccessor(requestedAt),
-      disableFilters: true,
-      disableSortBy: true
-    },
-    {
-      Header: 'Started At',
-      accessor: ({ startedAt }) => dateAccessor(startedAt),
-      disableFilters: true,
-      disableSortBy: true
-    },
-    {
-      Header: 'Finished At',
-      accessor: ({ finishedAt }) => dateAccessor(finishedAt),
-      disableFilters: true,
-      disableSortBy: true
-    },
-    {
-      Header: 'Output',
-      accessor: 'output',
-      disableFilters: true
-    }
-  ];
