--- frontend/src/pages/Organization/authContextOrg.tsx
@@ -1,9 +1,0 @@
-import { useAuthContext } from "context";
-
-export const { 
-    apiGet, 
-    apiPut, 
-    apiPost, 
-    user, 
-    setFeedbackMessage 
-} = useAuthContext();
